ref(tracing-internal): Export fetch instrumentation (#9473)
