+++ packages/tracing-internal/src/browser/index.ts
@@ -3,11 +3,7 @@ export * from '../exports';
 export type { RequestInstrumentationOptions } from './request';
 
 export { BrowserTracing, BROWSER_TRACING_INTEGRATION_ID } from './browsertracing';
-export {
-  instrumentOutgoingRequests,
-  defaultRequestInstrumentationOptions,
-  addTracingHeadersToFetchRequest,
-} from './request';
+export { instrumentOutgoingRequests, defaultRequestInstrumentationOptions } from './request';
 
 export {
   addPerformanceInstrumentationHandler,
