+++ packages/types/src/instrument.ts
@@ -31,6 +31,8 @@ interface SentryFetchData {
   url: string;
   request_body_size?: number;
   response_body_size?: number;
+  // span_id for the fetch request
+  __span?: string;
 }
 
 export interface HandlerDataFetch {
@@ -38,6 +40,18 @@ export interface HandlerDataFetch {
   fetchData: SentryFetchData;
   startTimestamp: number;
   endTimestamp?: number;
-  // This is actually `Response`, make sure to cast this where needed (not available in Node)
-  response?: unknown;
+  // This is actually `Response` - Note: this type is not complete. Add to it if necessary.
+  response?: {
+    readonly ok: boolean;
+    readonly status: number;
+    readonly url: string;
+    headers: {
+      append(name: string, value: string): void;
+      delete(name: string): void;
+      get(name: string): string | null;
+      has(name: string): boolean;
+      set(name: string, value: string): void;
+    };
+  };
+  error?: unknown;
 }
